requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
djangorestframework==3.14.0
djongo
dnspython==2.4.2
pymongo==4.5.0
pytz==2023.3.post1
sqlparse
typing-extensions==4.8.0
